pyproject.toml
[project]
name = "eisaku"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiomysql>=0.3.2",
    "alembic>=1.16.5",
    "fastapi-mail>=1.5.8",
    "fastapi[standard]>=0.116.1",
    "greenlet>=3.2.4",
    "jwt>=1.4.0",
    "passlib[bcrypt]>=1.7.4",
    "pymysql>=1.1.2",
    "python-decouple>=3.8",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
